Allow override of the number of parents at each level. Refs #19.
When a player or their parent is not able to cite details about their
one or both their parents, it should be possible to overload this in
such as way that they can proceed to declare.

This will not be a default, it will require an administrative action, so
the public user interface does not expose this.